Reorganize code moving it to different modules
The reusable parts of the ChangeLogMessage infra is moved
to admin_site to avoid circular imports
