+++ admin_site/viewsets.py
@@ -1,14 +1,18 @@
 from __future__ import annotations
 
-from typing import TYPE_CHECKING, Any
+from typing import TYPE_CHECKING, Any, Protocol
 
+from django.core.exceptions import PermissionDenied
+from django.db import models
 from django.db.models import Model, ProtectedError
 from django.forms.models import ModelForm
+from django.urls import reverse
 from django.utils.text import capfirst
-from django.utils.translation import gettext as _
+from django.utils.translation import gettext as _, pgettext_lazy
 from wagtail.admin import messages
 from wagtail.admin.forms.models import WagtailAdminModelForm
-from wagtail.snippets.views.snippets import CreateView, EditView, IndexView, SnippetViewSet
+from wagtail.admin.panels.field_panel import FieldPanel
+from wagtail.snippets.views.snippets import CreateView, DeleteView as SnippetDeleteView, EditView, IndexView, SnippetViewSet
 
 from kausal_common.users import user_or_bust
 
@@ -26,6 +30,12 @@ from admin_site.wagtail import execute_admin_post_save_tasks
 
 if TYPE_CHECKING:
     from django.http import HttpRequest
+
+    from actions.models.action import BaseChangeLogMessage
+
+
+class ObjectWithPublicChangeLogMessage(Protocol):
+    def get_public_change_log_message(self) -> BaseChangeLogMessage | None: ...
 
 
 class WatchEditView[ModelT: Model, FormT: WagtailAdminModelForm = WagtailAdminModelForm[Any]](
@@ -138,3 +148,121 @@ class WatchViewSet[ModelT: Model, FormT: ModelForm[Any] = WagtailAdminModelForm[
         if self._edit_handler and not self._edit_handler.base_form_class:
             self._edit_handler.base_form_class = WatchAdminModelForm[ModelT]  # type: ignore[assignment]
         return super().get_form_class(for_update)
+
+
+class BaseChangeLogMessageCreateView[M: models.Model, RelatedModel: ObjectWithPublicChangeLogMessage](WatchCreateView[M]):
+    related_field_name: str
+    success_url_name: str
+
+    def get_related_id(self) -> str | None:
+        """Get related object ID from GET params or POST data (hidden field)."""
+        if self.request.method == 'POST':
+            return self.request.POST.get(self.related_field_name)
+        return self.request.GET.get(self.related_field_name)
+
+    def get_related_object(self) -> RelatedModel | None:
+        related_id = self.get_related_id()
+        if not related_id:
+            return None
+        related_object = self.get_related_object_by_pk(related_id)
+        return related_object
+
+    def get_latest_change_log_message(self) -> BaseChangeLogMessage | None:
+        related_object = self.get_related_object()
+        if related_object is None:
+            return None
+        return related_object.get_public_change_log_message()
+
+    def get_related_object_by_pk(self, _pk: str) -> RelatedModel | None:
+        raise NotImplementedError
+
+    def check_related_object_permission(self, _related_obj: RelatedModel | None) -> bool:
+        raise NotImplementedError
+
+    def dispatch(self, request, *args, **kwargs):
+        related_obj = self.get_related_object()
+        if not self.check_related_object_permission(related_obj):
+            raise PermissionDenied
+        return super().dispatch(request, *args, **kwargs)
+
+    def get_page_subtitle(self):
+        related_obj = self.get_related_object()
+        if related_obj is not None:
+            return pgettext_lazy('page subtitle', 'Change history message: %(obj)s') % {'obj': related_obj}
+        return pgettext_lazy('page subtitle', 'Change history message')
+
+    def get_form(self, *args, **kwargs):
+        form = super().get_form(*args, **kwargs)
+        related_obj = self.get_related_object()
+        if related_obj is not None:
+            setattr(form.instance, self.related_field_name, related_obj)
+        form.instance.created_by = self.request.user  # type: ignore[attr-defined]
+        return form
+
+    def get_skip_url(self) -> str:
+        return reverse(self.success_url_name)
+
+    def get_context_data(self, **kwargs):
+        context = super().get_context_data(**kwargs)
+        context['skip_url'] = self.get_skip_url()
+        context['latest_change_log_message'] = self.get_latest_change_log_message()
+        context['related_field_name'] = self.related_field_name
+        context['related_id'] = self.get_related_id()
+        return context
+
+    def get_success_url(self):
+        return reverse(self.success_url_name)
+
+
+class BaseChangeLogMessageEditView[M: models.Model, RelatedModel: ObjectWithPublicChangeLogMessage](WatchEditView[M]):
+    related_field_name: str
+    success_url_name: str
+
+    def check_related_object_permission(self, _related_obj: RelatedModel | None) -> bool:
+        raise NotImplementedError
+
+    def dispatch(self, request, *args, **kwargs):
+        self.object = self.get_object()
+        related_obj = getattr(self.object, self.related_field_name, None)
+        if not self.check_related_object_permission(related_obj):
+            raise PermissionDenied
+        return super().dispatch(request, *args, **kwargs)
+
+    def get_success_url(self):
+        assert self.object is not None
+        related_obj = getattr(self.object, self.related_field_name)
+        return reverse(self.success_url_name, args=[related_obj.pk])
+
+
+class BaseChangeLogMessageDeleteView[M: models.Model, RelatedModel: ObjectWithPublicChangeLogMessage](SnippetDeleteView):
+    related_field_name: str
+
+    def check_related_object_permission(self, _related_obj: RelatedModel | None) -> bool:
+        raise NotImplementedError
+
+    def dispatch(self, request, *args, **kwargs):
+        self.object = self.get_object()
+        related_obj = getattr(self.object, self.related_field_name, None)
+        if not self.check_related_object_permission(related_obj):
+            raise PermissionDenied
+        return super().dispatch(request, *args, **kwargs)
+
+
+class BaseChangeLogMessageViewSet[M: models.Model](WatchViewSet[M]):
+    add_to_admin_menu = False
+    icon = 'doc-full'
+    page_title = pgettext_lazy('page title', 'Add change history message')
+    plan_filter_path: str
+    create_template_name = 'aplans/change_log_message_create.html'
+
+    panels = [
+        FieldPanel('content'),
+    ]
+
+    def get_queryset(self, request):
+        qs = super().get_queryset(request)
+        user = user_or_bust(request.user)
+        plan = user.get_active_admin_plan()
+        if qs is None:
+            return self.model.objects.none()  # type: ignore[attr-defined]
+        return qs.filter(**{self.plan_filter_path: plan})
